vtables -> templates
