+++ code/espurna/settings.h
@@ -17,6 +17,7 @@ Copyright (C) 2016-2019 by Xose Pérez <xose dot perez at gmail dot com>
 #include <ArduinoJson.h>
 
 #include "broker.h"
+#include "storage_eeprom.h"
 #include "settings_embedis.h"
 
 BrokerDeclare(ConfigBroker, void(const String& key, const String& value));
@@ -25,7 +26,27 @@ BrokerDeclare(ConfigBroker, void(const String& key, const String& value));
 
 namespace settings {
 
-extern embedis::KeyValueStore kv_store;
+struct EepromStorage {
+
+    uint8_t read(size_t pos) {
+        return EEPROMr.read(pos);
+    }
+
+    void write(size_t pos, uint8_t value) {
+        EEPROMr.write(pos, value);
+    }
+
+    void commit() {
+#if SETTINGS_AUTOSAVE
+        eepromCommit();
+#endif
+    }
+
+};
+
+using kvs_type = embedis::KeyValueStore<EepromStorage>;
+
+extern kvs_type kv_store;
 
 } // namespace settings
