+++ code/test/platformio.ini
@@ -9,7 +9,6 @@ test_build_project_src = true
 src_filter =
     +<../espurna/terminal_commands.cpp>
     +<../espurna/terminal_parsing.cpp>
-    +<../espurna/settings_embedis.cpp>
 lib_deps =
     StreamString
     https://github.com/bxparks/UnixHostDuino#d740398e
